refactors backend storage (does not compile)
