--- api/custom/storageapi.go
@@ -1,6 +1,0 @@
-package custom
-
-// StorageAPI specifies the interface that backend  engine needs to satisfy
-type StorageAPI interface {
-	CustomRequest(map[string]interface{}) (interface{}, error)
-}
