--- api/dbmeta/storageapi.go
@@ -1,8 +1,0 @@
-package dbmeta
-
-// StorageAPI specifies the interface that backend  engine needs to satisfy
-type StorageAPI interface {
-	GetVersion() (string, error)
-	GetDatabase() (string, string, error)
-	GetDatasets() (map[string]interface{}, error)
-}
