+++ api/npexplorer/storageapi.go
@@ -54,18 +54,3 @@ type CompletenessParams struct {
 	DatasetParams
 	FilterParams
 }
-
-// StorageAPI specifies the interface that backend engine needs to satisfy
-type StorageAPI interface {
-	ExplorerFindNeurons(FindNeuronsParams) (interface{}, error)
-	ExplorerNeuronMeta(DatasetParams) (interface{}, error)
-	ExplorerNeuronMetaVals(MetaValParams) (interface{}, error)
-	ExplorerROIConnectivity(DatasetParams) (interface{}, error)
-	ExplorerRankedTable(ConnectionsParams) (interface{}, error)
-	ExplorerSimpleConnections(ConnectionsParams) (interface{}, error)
-	ExplorerROIsInNeuron(NeuronNameParams) (interface{}, error)
-	ExplorerCommonConnectivity(CommonConnectivityParams) (interface{}, error)
-	ExplorerAutapses(DatasetParams) (interface{}, error)
-	ExplorerDistribution(DistributionParams) (interface{}, error)
-	ExplorerCompleteness(CompletenessParams) (interface{}, error)
-}
