+++ config/apis.go
@@ -3,3 +3,7 @@ package config
 import _ "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta"
 import _ "github.com/connectome-neuprint/neuPrintHTTP/api/custom"
 import _ "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer"
+import _ "github.com/connectome-neuprint/neuPrintHTTP/api/skeletons"
+import _ "github.com/connectome-neuprint/neuPrintHTTP/api/raw/cypher"
+import _ "github.com/connectome-neuprint/neuPrintHTTP/api/raw/spatial"
+import _ "github.com/connectome-neuprint/neuPrintHTTP/api/raw/keyvalue"
